some data type fixes for LP64 compatibility

+++ zaero/q_shared.h
@@ -177,7 +177,7 @@ extern vec3_t vec3_origin;
 #if defined(_MSC_VER) && defined(_M_IX86) && !defined(C_ONLY)
 extern long Q_ftol( float f );
 #else
-#define Q_ftol( f ) ( long ) (f)
+#define Q_ftol( f ) (int) (f)
 #endif
 
 #define DotProduct(x,y)			(x[0]*y[0]+x[1]*y[1]+x[2]*y[2])
